+++ manubot/cite/pubmed.py
@@ -6,7 +6,7 @@ import xml.etree.ElementTree
 
 import requests
 
-from .citekey import Handler
+from .handlers import Handler
 from manubot.util import get_manubot_user_agent
